pyproject.toml
[project]
name = "superyml"
version = "0.1.0"
description = "CLI tool to check and auto-fix YAML files recursively."
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "superyml contributors" }]
license = "MIT"
dependencies = [
  "ruamel.yaml>=0.18.6",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
superyml = "superyml.cli:main"

[build-system]
requires = ["setuptools>=68.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["superyml*"]
